requirements.txt
aiogram==2.25.1
psycopg2-binary==2.9.6
openai==0.27.6
tinytag==1.9.0
pandas==2.0.1
python-dotenv==1.0.0
python-unsplash==1.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
audioread==3.0.0
Babel==2.9.1
beets==1.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
confuse==2.0.1
contourpy==1.0.7
coverage==5.5
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
eyed3==0.9.7
fastapi==0.95.1
filetype==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
gino==1.0.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jellyfish==0.11.2
joblib==1.2.0
jsonpickle==3.0.1
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
log3==0.1.7
magic-filter==1.0.9
matplotlib==3.7.1
mediafile==0.11.0
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
munkres==1.1.4
music21==8.1.0
musicbrainzngs==0.7.1
mutagen==1.46.0
numba==0.57.0
numpy==1.24.3
oauthlib==2.0.1
openai==0.27.6
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pooch==1.6.0
psycopg2-binary==2.9.6
pyacoustid==1.2.2
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.30.0
requests-oauthlib==0.7.0
rich==12.6.0
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.22.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
spotify-dl==8.7.0
spotipy==2.23.0
SQLAlchemy==1.3.24
starlette==0.26.1
threadpoolctl==3.1.0
tinytag==1.9.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.15
uvicorn==0.22.0
webcolors==1.13
websockets==11.0.2
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2023.3.4
